# packages/backend/api/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Manna Financial Management Platform API
  description: |
    Comprehensive API for the Manna Financial Management Platform - an agentic accounting firm system.

    ## Features
    - Plaid API integration for bank account connectivity
    - ML-powered transaction categorization
    - Real-time financial reporting
    - JWT-based authentication
    - WebSocket support for real-time updates

    ## Authentication
    Most endpoints require JWT authentication via the `Authorization: Bearer <token>` header.

    ## Rate Limiting
    - Standard endpoints: 1000 requests/hour per user
    - Plaid sync endpoints: 100 requests/hour per user
    - ML categorization: 500 requests/hour per user

    ## WebSocket Events
    Connect to `/ws/updates` for real-time notifications of:
    - New transactions
    - Categorization updates
    - Report generation status
    - Account sync status

  version: 1.3.0
  contact:
    name: Manna Financial API Support
    email: api-support@manna.financial
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.manna.financial/v1
    description: Production server
  - url: https://staging-api.manna.financial/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: Get transactions
      description: Retrieve transactions with filtering, pagination, and sorting
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: account_id
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: min_amount
          in: query
          schema:
            type: number
        - name: max_amount
          in: query
          schema:
            type: number
        - name: is_business
          in: query
          schema:
            type: boolean
        - name: needs_review
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
            description: Search in merchant name or description
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [date, amount, merchant_name]
            default: date
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /accounts:
    get:
      tags:
        - Account Management
      summary: Get all accounts
      description: Retrieve all connected financial accounts
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    AccountBalance:
      type: object
      properties:
        available:
          type: number
          example: 1250.75
        current:
          type: number
          example: 1300.50
        limit:
          type: number
          nullable: true
          example: null
        iso_currency_code:
          type: string
          example: 'USD'
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          example: 'txn_12345'
        plaid_transaction_id:
          type: string
          example: 'plaid_txn_67890'
        account_id:
          type: string
          example: 'acc_12345'
        amount:
          type: number
          example: -25.50
          description: 'Negative for expenses, positive for income'
        date:
          type: string
          format: date
          example: '2024-01-01'
        authorized_date:
          type: string
          format: date
          nullable: true
          example: '2024-01-01'
        name:
          type: string
          example: 'STARBUCKS #1234'
        merchant_name:
          type: string
          nullable: true
          example: 'Starbucks'
        category:
          type: string
          nullable: true
          example: 'Food and Drink'
        subcategory:
          type: string
          nullable: true
          example: 'Coffee Shop'
        ml_category:
          type: string
          nullable: true
          example: 'Business Meals'
        ml_confidence:
          type: number
          nullable: true
          example: 0.85
        user_category:
          type: string
          nullable: true
          example: 'Business Meals'
        is_business:
          type: boolean
          example: true
        is_tax_deductible:
          type: boolean
          example: true
        pending:
          type: boolean
          example: false
        location:
          $ref: '#/components/schemas/TransactionLocation'
        payment_method:
          type: string
          nullable: true
          example: 'card'
        notes:
          type: string
          nullable: true
          example: 'Client meeting lunch'
        tags:
          type: array
          items:
            type: string
          example: ['client', 'deductible']
        created_at:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
    TransactionLocation:
      type: object
      properties:
        address:
          type: string
          nullable: true
          example: '123 Main St'
        city:
          type: string
          nullable: true
          example: 'San Francisco'
        state:
          type: string
          nullable: true
          example: 'CA'
        zip:
          type: string
          nullable: true
          example: '94102'
        country:
          type: string
          nullable: true
          example: 'US'
        lat:
          type: number
          nullable: true
          example: 37.7749
        lon:
          type: number
          nullable: true
          example: -122.4194
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          $ref: '#/components/schemas/Pagination'
        summary:
          $ref: '#/components/schemas/TransactionSummary'
    TransactionSummary:
      type: object
      properties:
        total_transactions:
          type: integer
          example: 1250
        total_amount:
          type: number
          example: -2450.75
        income_total:
          type: number
          example: 5000.00
        expense_total:
          type: number
          example: -7450.75
        by_category:
          type: object
          additionalProperties:
            type: number
          example:
            'Food and Drink': -450.25
            'Transportation': -125.50
            'Business Services': -1200.00
...
    Account:
      type: object
      properties:
        id:
          type: string
          example: 'acc_12345'
        plaid_account_id:
          type: string
          example: 'plaid_acc_67890'
        name:
          type: string
          example: 'Chase Business Checking'
        official_name:
          type: string
          example: 'Chase Premier Plus Business Checking'
        type:
          type: string
          enum: [depository, credit, loan, investment, other]
          example: 'depository'
        subtype:
          type: string
          example: 'checking'
        mask:
          type: string
          example: '0000'
        institution_name:
          type: string
          example: 'Chase'
        institution_id:
          type: string
          example: 'ins_3'
        balance:
          $ref: '#/components/schemas/AccountBalance'
        is_business:
          type: boolean
          example: true
        is_active:
          type: boolean
          example: true
        last_sync:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
        sync_status:
          type: string
          enum: [healthy, error, requires_update]
          example: 'healthy'
        created_at:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 50
        total_items:
          type: integer
          example: 1250
        total_pages:
          type: integer
          example: 25
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: 'Validation Error'
        message:
          type: string
          example: 'The provided data is invalid'
        code:
          type: string
          example: 'VALIDATION_ERROR'
        details:
          type: object
          additionalProperties: true
          example:
            field_errors:
              email: ['Invalid email format']
        timestamp:
          type: string
          format: date-time
          example: '2024-01-01T12:00:00Z'
        request_id:
          type: string
          example: 'req_12345'